drafting mlpro.sl.basics and FNN
