+++ src/mlpro/sl/__init__.py
@@ -1,1 +1,1 @@
-from mlpro.sl.models import *
+from mlpro.sl.basics import *
